Refactor compliantobs
